+++ ChangeLog.md
@@ -101,6 +101,8 @@
 
   * Fix docs for `examUuid` usage (Matt West).
 
+  * Fix `htmlparser2` config by copying default options from Cheerio (Nathan Walters).
+
   * Remove `allowIssueReporting` option in `infoCourseInstance.json` (Matt West).
 
   * Remove old temporary upgrade flag `tmp_upgraded_iq_status` (Matt West).